Creating Custom Particles guide overhaul
All changes are related to the Creating Custom Particles guide.
FEAT: Added things related to the sparkle particle, this includes sentence reworks and code reworks(including file renames)
FEAT: Added an image showcasing the custom particle
FEAT/REFACTOR: Added some tips and information boxes where I felt was fit. Some sections have been replaced to be a tip/info box instead of its own sentence
REFACTOR: Changed the particle's texture from the Fabric logo to a shiny blue sparkle
REFACTOR: Changed the Creating a JSON file and adding textures section to be more easily followed
FIX: JSON is now always properly capitalized when referencing it(not including the actual files)
FIX: Added a period on one of the code comments to be consistent with the rest of the code comments