Fix Rendering -> "Basic Rendering Concepts" page's mappings, fix the default VertexFormat list to match 1.21.10 (#436)
* Add self to authors

* Fix DefaultVertexFormat format lists:
Culled removed fields from lists;
Remapped to MojMaps

* FIx VertexFormat.Mode draw mode table format and mappings

* Remap multiply to rotateAround with param properly named as quaternionfc

* Fix typo "it's utility class" -> "its utility class"

* add translate(vec3) overload to PoseStack method list

* Remap "vertex" method to "addVertex" mojmap equivalent

* DrawMode.TRIANGLE_STRIP remapped to VertexFormat.Mode.TRIANGLE_STRIP
Added the outer class to remove confusion as Mode is a relatively generic sounding class name.

* Remove redundant "Mode." in the VertexFormat.Mode modes listing

* Remove guiGraphics.pose().last().pose() line of code which does not match actually defined methods
Someone more familiar with rendering may be more apt to replace with a proper code snippet, however it seems rather important to remove a line of code that is impossible, even if it means it currently has no replacement.